added db table skills and ability to show skills
