+++ src/main/resources/templates/profile.html
@@ -56,6 +56,18 @@
       <button type="submit" class="btn btn-primary">Delete your profile picture</button>
     </form>
   </div>
+  <div class ="container">
+    <h3>Your skills:</h3>
+    <span th:if="${skills.empty}">
+        <p>You haven't listed any skills!</p>
+    </span>
+    <span th:unless="${skills.empty}">
+      <ul th:each="skill: ${skills}">
+        <li th:text="${skill.name}">
+      </ul>
+      <p>Showing some skills</p>
+    </span>
+  </div>
 	<script src="https://code.jquery.com/jquery-3.4.0.js"
   	crossorigin="anonymous"></script>
 	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" crossorigin="anonymous"></script>
